FILE CODE
A Smart Luggage Carrier is a luggage carrier robot  which  follows  the  user automatically.  The main  aim of our project is to develop a Smart Luggage Carrier that could be user friendly. It will use wireless communication UART to communicate  with user. GPS is used for finding location of Master and location of Slave. Moreover, our project Smart Luggage Carrier is used in various applications such as Airport, Railway Station, etc. for carrying luggage.